requirements.txt
SQLAlchemy>=2.0
asyncpg
aiogram
python-dotenv